tests/wpt/tests/close-watcher/user-activation/n-destroy-n.html

<!doctype html>
<meta name=variant content="?dialog">
<meta name=variant content="?CloseWatcher">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/common/top-layer.js"></script>
<script src="../resources/helpers.js"></script>

<body>
<script>
const type = location.search.substring(1);

promise_test(async t => {
  const events = [];

  const watcher1 = createRecordingCloseWatcher(t, events, "watcher1", type);

  destroyCloseWatcher(watcher1);
  await waitForPotentialCloseEvent();
  assert_array_equals(events, []);

  createRecordingCloseWatcher(t, events, "watcher2", type);

  await sendCloseRequest();
  await waitForPotentialCloseEvent();
  assert_array_equals(events, ["watcher2 cancel[cancelable=false]", "watcher2 close"]);
}, "Create a close watcher without user activation; destroy the close watcher; create a close watcher without user activation");
</script>
